image_loader: Move image loading under "compose" namespace
Move the image loading functionality from the "docker" package to the
"compose" package because:

1. It does not make sense to have a separate package just for image
   loading in general.

2. It allows reusing the "compose" package functionality for the image
   loading without introducing another/third package to avoid circular
   imports.
