+++ pkg/compose/image_loader.go
@@ -1,4 +1,4 @@
-package docker
+package compose
 
 import (
 	"archive/tar"
