+++ pkg/compose/image_loader_test.go
@@ -1,4 +1,4 @@
-package docker
+package compose
 
 import "testing"
