+++ pkg/compose/install.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"github.com/docker/docker/pkg/archive"
 	"github.com/foundriesio/composeapp/internal/progress"
-	"github.com/foundriesio/composeapp/pkg/docker"
 	"os"
 	"path"
 )
@@ -16,7 +15,7 @@ type (
 	InstallProgress struct {
 		AppInstallState AppInstallState
 		AppID           string
-		ImageLoadState  docker.ImageLoadState
+		ImageLoadState  ImageLoadState
 		ImageID         string
 		ID              string
 		Current         int64
@@ -91,7 +90,7 @@ func Install(ctx context.Context,
 		return err
 	}
 
-	appImageURIs := make(docker.ImageDescriptions)
+	appImageURIs := make(ImageDescriptions)
 	err = app.GetComposeRoot().Walk(func(node *TreeNode, depth int) error {
 		if node.Type == BlobTypeImageManifest {
 			nodeURI := node.Descriptor.URLs[0]
@@ -103,9 +102,9 @@ func Install(ctx context.Context,
 		return err
 	}
 
-	err = docker.LoadImages(ctx, cli, appImageURIs, blobsRoot, docker.WithRefWithDigest(),
-		docker.WithBlobReadingFromStore(),
-		docker.WithProgressReporting(func(ilp *docker.LoadImageProgress) {
+	err = LoadImages(ctx, cli, appImageURIs, blobsRoot, WithRefWithDigest(),
+		WithBlobReadingFromStore(),
+		WithProgressReporting(func(ilp *LoadImageProgress) {
 			if opts.ProgressReporter != nil {
 				opts.ProgressReporter.Update(InstallProgress{
 					AppInstallState: AppInstallStateImagesLoading,
@@ -120,7 +119,19 @@ func Install(ctx context.Context,
 		}))
 	if err != nil {
 		// Try to load images without pinning to digest and reading directly from the store
-		err = docker.LoadImages(ctx, cli, appImageURIs, blobsRoot)
+		err = LoadImages(ctx, cli, appImageURIs, blobsRoot, WithProgressReporting(func(ilp *LoadImageProgress) {
+			if opts.ProgressReporter != nil {
+				opts.ProgressReporter.Update(InstallProgress{
+					AppInstallState: AppInstallStateImagesLoading,
+					AppID:           app.Ref().String(),
+					ImageLoadState:  ilp.State,
+					ImageID:         ilp.ImageID,
+					ID:              ilp.ID,
+					Current:         ilp.Current,
+					Total:           ilp.Total,
+				})
+			}
+		}))
 	}
 	return err
 }
