+++ pkg/compose/tar_streamer.go
@@ -1,4 +1,4 @@
-package docker
+package compose
 
 import (
 	"archive/tar"
