+++ pkg/compose/tar_streamer_test.go
@@ -1,4 +1,4 @@
-package docker
+package compose
 
 import (
 	"archive/tar"
